Logic for making new grid done
Logic for making new grid in the same div's place is done. Added formula
for calculating the grid size and the size of its subelements div's.
Conditioned the user to enter only numbers from 1 to 100. Added new
buttons for the rgb colors to be added and added event listener to it to
wait for the mouseevent when the button is clicked. The same is done for
the erase buttons. One is erasing everything and the other is erasing
one square at the time.